Refactor out custom columns
